requirements.txt
griptape[all]@git+https://github.com/griptape-ai/griptape@french/240514/transcription--transcription
requests
python-dotenv
elevenlabs
